requirements.txt
Django>=4.2,<6.0
Pillow>=10.0
python-dotenv>=1.0
google-genai>=0.1
